refactor(email settings): extract preview email layout
